ai/LLMApi/requirements.txt
# Azure Functions runtime
azure-functions==1.20.0

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1

# Azure services
azure-storage-blob==12.19.0
azure-servicebus==7.11.4
azure-identity==1.15.0

# OpenAI
openai==2.14.0
httpx==0.27.2  

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Data validation
pydantic>=2.7.0
pydantic-settings==2.5.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
